Refactor scraping logic to a separate function
Refactored the logic for scraping meta data into a separate function called getMetaDataForEntryDataList. This was done to remove repetitive code blocks and promote code reusability. We can now easily scrape meta data across multiple files - app/page.tsx, app/medium/[slug]/page.tsx, app/tag/[slug]/page.tsx, and app/entry/[slug]/page.tsx - by calling this new function.
